+++ packages/core-cairo/src/generate/erc20.ts
@@ -13,6 +13,9 @@ const blueprint = {
   pausable: booleans,
   mintable: booleans,
   premint: ['1'],
+  votes: booleans,
+  appName: ['MyApp'],
+  appVersion: ['v1'],
   access: accessOptions,
   upgradeable: upgradeableOptions,
   info: infoOptions