+++ packages/core-cairo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.12.0 (2024-05-01)
+
+- Add `votes` option to ERC20. ([#355](https://github.com/OpenZeppelin/contracts-wizard/pull/355))
+- Require `appName` and `appVersion` when `votes` is enabled.
+- Sort implemented traits.
+- **Breaking changes**:
+  - Use OpenZeppelin Contracts for Cairo v0.12.0.
+  - Use Hooks with ERC20.
+
 ## 0.11.0 (2024-04-17)
 
 - **Breaking changes**: